Add copy_gate_5500 and copy_gate_50M evaluation results
copy_gate_5500 (0.72B tier, gate on baseline_10x_less):
- val_loss: 3.4151, Story QA: 28%, p_copy: 4.1%
- Checkpoint: checkpoints/copy_gate_5500.pt (364 MB, on disk)

copy_gate_50M (0.05B tier, gate on baseline_50M_retrain):
- val_loss: 4.2903, Story QA: 1%, p_copy: 5.4%
- Checkpoint: checkpoints/copy_gate_50M.pt (436 MB, on disk)

Bug fixes:
- model.py: Fix scatter_add_ dtype mismatch under bf16 autocast
- experiment_tracker.py: Add copy_gate to model choices

+++ model.py
@@ -397,7 +397,7 @@ class CopyGateTransformer(SimpleTransformer):
         copy_probs.scatter_add_(
             2,
             input_ids.unsqueeze(1).expand(-1, T, -1),
-            copy_attn,
+            copy_attn.to(copy_probs.dtype),
         )
 
         # Blend generation and copy distributions
